execute submit tx through client
